requirements.txt
certifi==2020.6.20
cffi==1.14.1
click==7.1.2
cryptography==3.0
decorator==4.4.2
dnspython==1.16.0
future==0.18.2
idna==2.10
numpy==1.19.1
pycparser==2.20
pymongo==3.11.0
PySocks==1.7.1
python-dotenv==0.14.0
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
ujson==3.0.0
pytz